create posting page(have not been finished)
