--- app/src/main/java/com/example/skr/ui/fan.java
@@ -1,22 +1,0 @@
-package com.example.skr.ui;
-
-public class fan {
-    private String userAccount;//被关注的人的账户
-    private  String  fan;//点关注的人的账户
-
-    public String getUserAccount() {
-        return userAccount;
-    }
-
-    public String getFan() {
-        return fan;
-    }
-
-    public void setUserAccount(String userAccount) {
-        this.userAccount = userAccount;
-    }
-
-    public void setFan(String fan) {
-        this.fan = fan;
-    }
-}
